Correct checkstyle errors.
