+++ data/saved.txt
@@ -1,1 +1,1 @@
-T | 0 | return bookT | 0 | return bookT | 0 | return bookT | 0 | return book
+T | 0 | return bookT | 0 | return bookT | 0 | return bookT | 0 | return bookT | 0 | return bookT | 0 | return bookT | 0 | return bookT | 0 | return bookT | 0 | borrow bookT | 1 | return bookT | 1 | borrow bookT | 0 | return bookT | 1 | return bookT | 0 | return bookT | 0 | return bookT | 1 | return bookT | 0 | return bookT | 0 | return bookT | 1 | return bookD | 0 | borrow book 2020-02-02T18:00:00
