--- src/main/java/Ally/Commands/Commands.java
@@ -1,10 +1,0 @@
-package Ally.Commands;
-
-import Ally.Exceptions.AllyException;
-import Ally.Tasks.AllyList;
-import Ally.Storage;
-import Ally.Ui;
-public abstract class Commands {
-    public abstract void run (AllyList allyList, Ui ui, Storage storage) throws AllyException;
-    public abstract boolean isExit();
-}
