+++ src/main/java/ally/commands/ListCommand.java
@@ -1,10 +1,13 @@
-package Ally.Commands;
+package ally.commands;
 
-import Ally.Exceptions.AllyException;
-import Ally.Tasks.AllyList;
-import Ally.Storage;
-import Ally.Ui;
+import ally.Storage;
+import ally.Ui;
+import ally.exceptions.AllyException;
+import ally.tasks.AllyList;
 
+/**
+ * ListCommand inherits from Commands.
+ */
 public class ListCommand extends Commands {
 
     /**
