+++ src/main/java/ally/commands/UnmarkCommand.java
@@ -1,26 +1,28 @@
-package Ally.Commands;
+package ally.commands;
 
-import Ally.*;
-import Ally.Exceptions.AllyException;
-import Ally.Tasks.AllyList;
-import Ally.Tasks.Task;
+import ally.Storage;
+import ally.Ui;
+import ally.exceptions.AllyException;
+import ally.tasks.AllyList;
+import ally.tasks.Task;
 
-public class MarkCommand extends Commands {
-
+/**
+ * UnmarkCommand inherits from Commands.
+ */
+public class UnmarkCommand extends Commands {
     private final int index;
 
     /**
-     * Constructor for MarkCommand.
+     * Constructor for UnmarkCommand.
      *
      * @param index
      */
-    public MarkCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
     /**
-     * Marks the task as done and save the marked task to the saved file.
-     *
+     * Marks the task as not done and saved the task to the saved file.
      * @param allyList
      * @param ui
      * @param storage
@@ -29,7 +31,8 @@ public class MarkCommand extends Commands {
     public void run(AllyList allyList, Ui ui, Storage storage) {
         try {
             Task task = allyList.getTask(index);
-            ui.showMarked(task);
+            allyList.unMarkDone(index);
+            ui.showNotMarked(task);
             storage.appendToFile(task);
         } catch (AllyException e) {
             System.out.println(e.getMessage());
@@ -40,5 +43,4 @@ public class MarkCommand extends Commands {
     public boolean isExit() {
         return false;
     }
-
 }
