+++ src/main/java/ally/exceptions/AllyException.java
@@ -1,5 +1,8 @@
-package Ally.Exceptions;
+package ally.exceptions;
 
+/**
+ * Exception Class for Ally.
+ */
 public class AllyException extends Exception {
     public AllyException(String message) {
         super(message);
