+++ src/test/java/ally/tasks/DeadlineTest.java
@@ -1,41 +1,43 @@
-package Ally.Tasks;
+package ally.tasks;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 public class DeadlineTest {
 
     @Test
-    public void getStatusIcon_markAsDone_success(){
+    public void getStatusIcon_markAsDone_success() {
         Deadline deadline = new Deadline("Sleep", "by 2020-02-02 1800");
         deadline.setMarkDone();
         assertEquals("X", deadline.getStatusIcon());
     }
 
     @Test
-    public void getStatusIcon_markAsNotDone_success(){
+    public void getStatusIcon_markAsNotDone_success() {
         Deadline deadline = new Deadline("Sleep", "by 2020-02-02 1800");
         deadline.setMarkNotDone();
         assertEquals(" ", deadline.getStatusIcon());
     }
 
     @Test
-    public void markAsDone_success(){
+    public void markAsDone_success() {
         Deadline deadline = new Deadline("Sleep", "by 2020-02-02 1800");
         deadline.setMarkDone();
         assertTrue(deadline.isDone);
     }
 
     @Test
-    public void markAsNotDone_success(){
+    public void markAsNotDone_success() {
         Deadline deadline = new Deadline("Sleep", "by 2020-02-02 1800");
         deadline.setMarkNotDone();
         assertFalse(deadline.isDone);
     }
 
     @Test
-    public void formatFile_markAsUndone_success(){
+    public void formatFile_markAsUndone_success() {
         Deadline deadline = new Deadline("Sleep", "by 2020-02-02 1800");
         deadline.setMarkNotDone();
         String expected = "D | 0 | Sleep 2020-02-02T18:00:00";
@@ -43,7 +45,7 @@ public class DeadlineTest {
     }
 
     @Test
-    public void formatFile_markAsDone_success(){
+    public void formatFile_markAsDone_success() {
         Deadline deadline = new Deadline("Sleep", "by 2020-02-02 1800");
         deadline.setMarkDone();
         String expected = "D | 1 | Sleep 2020-02-02T18:00:00";
@@ -51,7 +53,7 @@ public class DeadlineTest {
     }
 
     @Test
-    public void toString_markAsUndone_success(){
+    public void toString_markAsUndone_success() {
         Deadline deadline = new Deadline("Sleep", "by 2020-02-02 1800");
         deadline.setMarkNotDone();
         String expected = "[D][ ] Sleep (by: Feb 02 2020 6pm)";
@@ -59,7 +61,7 @@ public class DeadlineTest {
     }
 
     @Test
-    public void toString_markAsDone_success(){
+    public void toString_markAsDone_success() {
         Deadline deadline = new Deadline("Sleep", "by 2020-02-02 1800");
         deadline.setMarkDone();
         String expected = "[D][X] Sleep (by: Feb 02 2020 6pm)";
